♻️ refactor : 클래스 내부 메서드 이동
각각 클래스 내부의 메서드 중 클래스와 어울리지 않는 메서드를 utils.js 로 이동.
- pickRandomNumber()
- tryNumvalidation()
- createBoard()
