+++ src/App.js
@@ -3,6 +3,7 @@ import ioMethod from './inputoutputMethod.js';
 import MAGICNUMBER from './magicnumber.js';
 import ERRORMESAGE from './errorMessage.js';
 import validation from './validation.js';
+import utils from './utils.js';
 
 class App {
   #racingcar;
@@ -20,28 +21,9 @@ class App {
     const inputArr = inputs.split(MAGICNUMBER.SEPARATOR);
     this.#racingcar = new Racingcar(inputArr);
     this.#tryNum = await ioMethod.inputTryNum();
-    this.#tryNumvalidation(this.#tryNum);
-    this.#board = this.#createBoard(inputArr);
+    this.#tryNum = utils.tryNumvalidation(this.#tryNum);
+    this.#board = utils.createBoard(inputArr);
     ioMethod.printProgressResult();
-  }
-
-  #tryNumvalidation(tryNum) {
-    if (!validation.isPositiveInteger(Number(tryNum)))
-      throw new Error(`${ERRORMESAGE.NOT_POSITIVE_INTEGER}`);
-
-    const tryNumArr = String(tryNum).split('');
-    if (validation.hasSpace(tryNumArr))
-      throw new Error(`${ERRORMESAGE.WRONG_NUMBER_FORM_SPACE}`);
-
-    Number(this.#tryNum);
-  }
-
-  #createBoard(inputArr) {
-    const board = {};
-    inputArr.map((car) => {
-      board[car] = '';
-    });
-    return board;
   }
 
   #carRacing() {
