--- src/pickRandomNumber.js
@@ -1,11 +1,0 @@
-import MAGICNUMBER from './magicnumber.js';
-import { MissionUtils } from '@woowacourse/mission-utils';
-
-const pickRandomNumber = () => {
-  return MissionUtils.Random.pickNumberInRange(
-    MAGICNUMBER.STARTINCLUSIVE,
-    MAGICNUMBER.ENDINCLUSIVE,
-  );
-};
-
-export default pickRandomNumber;
